changed code for getting 'duration_sec' so the value won't be rounded anymore

+++ code/data/download_videos.py
@@ -80,7 +80,7 @@ def get_video_metadata(video_path):
         "fps": cap.get(cv2.CAP_PROP_FPS),
         "width": int(cap.get(cv2.CAP_PROP_FRAME_WIDTH)),
         "height": int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT)),
-        "duration_sec": int(cap.get(cv2.CAP_PROP_FRAME_COUNT) / cap.get(cv2.CAP_PROP_FPS)) if cap.get(cv2.CAP_PROP_FPS) > 0 else 0
+        "duration_sec": cap.get(cv2.CAP_PROP_FRAME_COUNT) / cap.get(cv2.CAP_PROP_FPS) if cap.get(cv2.CAP_PROP_FPS) > 0 else 0
     }
     
     cap.release()
